[chore] : 양의 소수도 입력되도록 처리
BigDecimal로 매우 큰 실수에 대해 처리할 수 있도록 변경

+++ src/main/java/calculator/controller/CalculatorController.java
@@ -4,6 +4,7 @@ import calculator.model.Calculator;
 import calculator.model.CalculatorImpl;
 import calculator.utils.InputValidator;
 import calculator.view.OutputView;
+import java.math.BigDecimal;
 
 public class CalculatorController {
 
@@ -21,7 +22,7 @@ public class CalculatorController {
         calculator.insertValues(input);
         calculator.extract();
 
-        int result = calculator.sum();
+        BigDecimal result = calculator.sum();
         OutputView.getOutputMessage(result);
     }
