+++ src/main/java/calculator/model/Calculator.java
@@ -1,7 +1,9 @@
 package calculator.model;
 
+import java.math.BigDecimal;
+
 public interface Calculator {
     void insertValues(String input);
     void extract();
-    int sum();
+    BigDecimal sum();
 }
